# contracts/openapi-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Theeb Engine Client API",
    "version": "1.0.0",
    "description": "Stable content-centric contract. Provider IDs, source URLs and scraper details are intentionally excluded."
  },
  "servers": [{ "url": "/v1" }],
  "paths": {
    "/search": {
      "get": {
        "operationId": "search",
        "parameters": [{ "name": "q", "in": "query", "required": true, "schema": { "type": "string", "minLength": 1, "maxLength": 200 } }],
        "responses": { "200": { "$ref": "#/components/responses/SearchResults" }, "400": { "$ref": "#/components/responses/Error" } }
      }
    },
    "/series/{id}": {
      "get": {
        "operationId": "getSeries",
        "parameters": [{ "$ref": "#/components/parameters/Id" }],
        "responses": { "200": { "$ref": "#/components/responses/Series" }, "404": { "$ref": "#/components/responses/Error" } }
      }
    },
    "/series/{id}/episodes": {
      "get": {
        "operationId": "listEpisodes",
        "parameters": [{ "$ref": "#/components/parameters/Id" }],
        "responses": { "200": { "$ref": "#/components/responses/EpisodeList" }, "404": { "$ref": "#/components/responses/Error" } }
      }
    },
    "/episodes/{id}": {
      "get": {
        "operationId": "getEpisode",
        "parameters": [{ "$ref": "#/components/parameters/Id" }],
        "responses": { "200": { "$ref": "#/components/responses/Episode" }, "404": { "$ref": "#/components/responses/Error" } }
      }
    },
    "/playback/sessions": {
      "post": {
        "operationId": "createPlaybackSession",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreatePlaybackSession" } } } },
        "responses": { "201": { "$ref": "#/components/responses/PlaybackSession" }, "400": { "$ref": "#/components/responses/Error" }, "404": { "$ref": "#/components/responses/Error" } }
      }
    },
    "/playback/sessions/{id}": {
      "get": {
        "operationId": "getPlaybackSession",
        "parameters": [{ "$ref": "#/components/parameters/SessionId" }],
        "responses": { "200": { "$ref": "#/components/responses/PlaybackSession" }, "404": { "$ref": "#/components/responses/Error" } }
      }
    },
    "/playback/sessions/{id}/feedback": {
      "post": {
        "operationId": "sendPlaybackFeedback",
        "parameters": [{ "$ref": "#/components/parameters/SessionId" }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PlaybackFeedback" } } } },
        "responses": { "202": { "$ref": "#/components/responses/FeedbackReceipt" }, "400": { "$ref": "#/components/responses/Error" }, "404": { "$ref": "#/components/responses/Error" }, "409": { "$ref": "#/components/responses/Error" }, "429": { "$ref": "#/components/responses/Error" } }
      }
    },
    "/episodes/{id}/download-options": {
      "get": {
        "operationId": "listDownloadOptions",
        "description": "Separate from playback. Listing never starts a download and never returns a temporary provider URL.",
        "parameters": [{ "$ref": "#/components/parameters/Id" }],
        "responses": { "200": { "$ref": "#/components/responses/DownloadOptions" }, "404": { "$ref": "#/components/responses/Error" } }
      }
    }
  },
  "components": {
    "parameters": {
      "Id": { "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "minimum": 1 } },
      "SessionId": { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
    },
    "schemas": {
      "Series": {
        "type": "object", "additionalProperties": false, "required": ["id", "title", "content_type", "status", "episode_count"],
        "properties": { "id": { "type": "integer" }, "title": { "type": "string" }, "original_title": { "type": ["string", "null"] }, "description": { "type": ["string", "null"] }, "image": { "type": ["string", "null"] }, "content_type": { "type": "string" }, "language": { "type": ["string", "null"] }, "country": { "type": ["string", "null"] }, "year": { "type": ["string", "null"] }, "status": { "type": "string" }, "episode_count": { "type": "integer", "minimum": 0 } }
      },
      "Episode": {
        "type": "object", "additionalProperties": false, "required": ["id", "canonical_series_id", "season_number", "watch_available", "download_available"],
        "properties": { "id": { "type": "integer" }, "canonical_series_id": { "type": "integer" }, "season_number": { "type": "integer" }, "episode_number": { "type": ["integer", "null"] }, "title": { "type": ["string", "null"] }, "description": { "type": ["string", "null"] }, "image": { "type": ["string", "null"] }, "watch_available": { "type": "boolean" }, "download_available": { "type": "boolean" } }
      },
      "CreatePlaybackSession": {
        "type": "object", "additionalProperties": false, "required": ["canonical_episode_id", "client"],
        "properties": {
          "canonical_episode_id": { "type": "integer", "minimum": 1 },
          "quality": { "enum": ["auto", "1080p", "720p", "480p"], "default": "auto" },
          "client": { "type": "object", "additionalProperties": false, "required": ["platform"], "properties": { "platform": { "enum": ["android", "android_tv", "ios", "web", "windows"] }, "version": { "type": "string", "maxLength": 50 } } }
        }
      },
      "PlaybackSession": {
        "type": "object", "additionalProperties": false, "required": ["id", "canonical_episode_id", "state", "requested_quality", "client", "plan_version", "playback", "created_at", "updated_at", "expires_at"],
        "properties": {
          "id": { "type": "string", "format": "uuid" }, "canonical_episode_id": { "type": "integer" },
          "state": { "enum": ["planning", "ready", "unavailable", "cancelled", "expired"] },
          "requested_quality": { "type": "string" }, "client": { "type": "object", "additionalProperties": false, "required": ["platform"], "properties": { "platform": { "enum": ["android", "android_tv", "ios", "web", "windows"] }, "version": { "type": ["string", "null"] } } }, "plan_version": { "type": "integer" },
          "playback": { "oneOf": [{ "type": "null" }, { "type": "object", "additionalProperties": false, "required": ["uri", "quality"], "properties": { "uri": { "type": "string", "pattern": "^/v1/playback/sessions/" }, "quality": { "type": "string" } } }] },
          "created_at": { "type": "string", "format": "date-time" }, "updated_at": { "type": "string", "format": "date-time" }, "expires_at": { "type": "string", "format": "date-time" }
        }
      },
      "PlaybackFeedback": {
        "type": "object", "additionalProperties": false, "required": ["event_id", "type", "occurred_at"],
        "properties": {
          "event_id": { "type": "string", "maxLength": 100 },
          "type": { "enum": ["player_opened", "first_frame", "playing", "buffering", "stalled", "ended", "fatal_error"] },
          "occurred_at": { "type": "string", "format": "date-time", "description": "Client timestamp. Values outside the server clock-skew window are normalized to server receive time." },
          "position_seconds": { "type": "number", "minimum": 0 },
          "error_code": { "type": "string", "maxLength": 100, "pattern": "^[A-Z0-9_.-]+$" },
          "details": {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "reason": { "type": "string", "maxLength": 200 },
              "player_state": { "type": "string", "maxLength": 200 },
              "network_type": { "type": "string", "maxLength": 200 },
              "buffered_seconds": { "type": "number" },
              "duration_seconds": { "type": "number" },
              "transport": { "type": "string", "maxLength": 200 }
            }
          }
        }
      },
      "DownloadOption": {
        "type": "object", "additionalProperties": false, "required": ["id", "status"],
        "properties": { "id": { "type": "string" }, "quality": { "type": ["string", "null"] }, "format": { "type": ["string", "null"] }, "status": { "enum": ["resolvable", "available", "unavailable"] }, "type": { "type": ["string", "null"] } }
      },
      "FeedbackReceipt": { "type": "object", "additionalProperties": false, "required": ["accepted", "duplicate"], "properties": { "accepted": { "const": true }, "duplicate": { "type": "boolean" } } },
      "Error": { "type": "object", "additionalProperties": false, "required": ["code", "message"], "properties": { "code": { "type": "string" }, "message": { "type": "string" } } }
    },
    "responses": {
      "SearchResults": { "description": "Canonical search results", "content": { "application/json": { "schema": { "type": "object", "additionalProperties": false, "required": ["data"], "properties": { "data": { "type": "object", "additionalProperties": false, "required": ["query", "count", "items"], "properties": { "query": { "type": "string" }, "count": { "type": "integer" }, "items": { "type": "array", "items": { "$ref": "#/components/schemas/Series" } } } } } } } } },
      "Series": { "description": "Canonical series", "content": { "application/json": { "schema": { "type": "object", "properties": { "data": { "$ref": "#/components/schemas/Series" } } } } } },
      "Episode": { "description": "Canonical episode", "content": { "application/json": { "schema": { "type": "object", "properties": { "data": { "$ref": "#/components/schemas/Episode" } } } } } },
      "EpisodeList": { "description": "Episodes", "content": { "application/json": { "schema": { "type": "object", "properties": { "data": { "type": "object", "properties": { "count": { "type": "integer" }, "items": { "type": "array", "items": { "$ref": "#/components/schemas/Episode" } } } } } } } } },
      "PlaybackSession": { "description": "Playback session", "content": { "application/json": { "schema": { "type": "object", "properties": { "data": { "$ref": "#/components/schemas/PlaybackSession" } } } } } },
      "FeedbackReceipt": { "description": "Accepted, including idempotent duplicates", "content": { "application/json": { "schema": { "type": "object", "additionalProperties": false, "required": ["data"], "properties": { "data": { "$ref": "#/components/schemas/FeedbackReceipt" } } } } } },
      "DownloadOptions": { "description": "User-selectable download options", "content": { "application/json": { "schema": { "type": "object", "properties": { "data": { "type": "object", "required": ["canonical_episode_id", "count", "items", "automatic_download", "action_required"], "properties": { "canonical_episode_id": { "type": "integer" }, "count": { "type": "integer" }, "items": { "type": "array", "items": { "$ref": "#/components/schemas/DownloadOption" } }, "automatic_download": { "const": false }, "action_required": { "const": "user_selection" } } } } } } } },
      "Error": { "description": "Stable error envelope", "content": { "application/json": { "schema": { "type": "object", "properties": { "error": { "$ref": "#/components/schemas/Error" } } } } } }
    }
  }
}